Refactor directory structure and adjust module imports
